chat-backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pytest-django
gunicorn